+++ cmd/login/error.go
@@ -190,3 +190,21 @@ var releaseNotFoundError = &microerror.Error{
 func IsReleaseNotFound(err error) bool {
 	return microerror.Cause(err) == releaseNotFoundError
 }
+
+var noOrganizationsError = &microerror.Error{
+	Kind: "noOrganizationsError",
+}
+
+// IsNoOrganizations asserts noOrganizationsError.
+func IsNoOrganizations(err error) bool {
+	return microerror.Cause(err) == noOrganizationsError
+}
+
+var multipleClustersFoundError = &microerror.Error{
+	Kind: "multipleClustersFoundError",
+}
+
+// IsMultipleClustersFound asserts multipleClustersFoundError.
+func IsMultipleClustersFound(err error) bool {
+	return microerror.Cause(err) == multipleClustersFoundError
+}