+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 - Usa CAPI templates for all releases from `v20.0.0-alpha1` onwards, to include alpha and beta releases.
 - Move AWS Cluster templating from `apiextensions`
 - Move AWS Node Pool templating from `apiextensions`
+- Make the `--organization` flag optional when using the `login` command with a workload cluster. The cluster will be searched in all the organization namespaces that the user has access to.
 
 ## [1.45.0] - 2021-10-26